Reorganize code to eliminate many of the ifdef HW_XFT
