+++ server/hw/hw_x/x11_alloccolor.c
