+++ server/hw/hw_x/x11_quitflavor.c
@@ -1,0 +1,4 @@
+/* Does the "flavor-specific" X11_QuitHW processing.  This one is for the basic X11 flavor. */
+static void X11_QuitFlavor(void) {
+    if (xsfont) XFreeFont(xdisplay, xsfont);
+}
