+++ server/hw/hw_x/x11_utf32_to_charset.c
@@ -1,3 +1,4 @@
+/* basic X-11 version of X11_UTF_32_to_charset_function */
 static Tutf_function X11_UTF_32_to_charset_function(CONST byte *charset) {
     XFontProp *fp;
     unsigned long prop;
