+++ server/hw/hw_xft/x11_alloccolor.c
