+++ server/hw/hw_xft/x11_setcolors.c
